添加登录验证

+++ src/main/java/cn/edu/upc/yb/integrate/homepage/controller/AppController.java
@@ -2,6 +2,8 @@ package cn.edu.upc.yb.integrate.homepage.controller;
 
 
 import cn.edu.upc.yb.integrate.calendar.dto.JsonMes;
+import cn.edu.upc.yb.integrate.common.model.CommonAdmin;
+import cn.edu.upc.yb.integrate.common.service.CommonAdminService;
 import cn.edu.upc.yb.integrate.homepage.model.App;
 import cn.edu.upc.yb.integrate.homepage.repository.AppRepository;
 import cn.edu.upc.yb.integrate.homepage.storage.StorageFileNotFoundException;
@@ -29,6 +31,9 @@ public class AppController {
     @Autowired
     public AppRepository appRepository;
 
+    @Autowired
+    public CommonAdminService commonAdminService;
+
     @GetMapping("/icon/{filename:.+}")
     @ResponseBody
     public ResponseEntity<Resource> showFile(@PathVariable String filename) {
@@ -41,39 +46,49 @@ public class AppController {
     }
 
     @GetMapping("/showall")
-    public Iterable<App> showAll(){
-        return appRepository.findAll();
+    public Object showAll(){
+            return appRepository.findAll();
     }
 
     @GetMapping("/showbytab")
     public Object showbytab(Integer tabid){
-        return appRepository.findByTabid(tabid);
+            return appRepository.findByTabid(tabid);
     }
 
     @PostMapping("/create")
     public Object create(@RequestParam("icon") MultipartFile icon, String name ,Integer tabid, String href) {
-        storageService.store(icon,name);
-        App app = new App(name,tabid,href);
-        appRepository.save(app);
-        return new JsonMes(0,"添加成功");
+        if(commonAdminService.isCommonAdmin()){
+            storageService.store(icon,name);
+            App app = new App(name,tabid,href);
+            appRepository.save(app);
+            return new JsonMes(0,"添加成功");
+        }else{
+            return new JsonMes(1,"你还没有登录或者不是管理员");
+        }
+
     }
     @PostMapping("/update")
     public Object update(Integer id, @RequestParam("file") MultipartFile icon, String name,Integer tabid, String href){
-        storageService.store(icon,name);
-        App app = appRepository.findOne(id);
-        app.update(name,tabid,href);
-        appRepository.save(app);
-        return new JsonMes(0,"更新成功");
+        if(commonAdminService.isCommonAdmin()){
+            storageService.store(icon,name);
+            App app = appRepository.findOne(id);
+            app.update(name,tabid,href);
+            appRepository.save(app);
+            return new JsonMes(0,"更新成功");
+        }else{
+            return new JsonMes(1,"你还没有登录或者不是管理员");
+        }
     }
-    @GetMapping("/update")
+    @GetMapping("/delete")
     public Object delete(Integer id){
-        appRepository.delete(id);
-        return new JsonMes(0,"删除成功");
+        if(commonAdminService.isCommonAdmin()){
+            appRepository.delete(id);
+            return new JsonMes(0,"删除成功");
+        }else{
+            return new JsonMes(1,"你还没有登录或者不是管理员");
+        }
     }
-    @GetMapping("/findone")
-    public Object find(Integer id){
-        return  appRepository.findOne(id);
-    }
+
